requirements.txt
requests==2.20.0
beautifulsoup4==4.6.3
tweepy==3.6.0
Mastodon.py==1.3.1
